refactor(idl): remove the schema shim and split codegen internals
Keep naming private to its consumers, separate TypeScript package/codec/PDA generation, and resolve every quasar-svm consumer from the published registry source.
