+++ idl/Cargo.toml
@@ -24,7 +24,7 @@ doctest = false
 
 [dependencies]
 bs58 = "0.5.1"
-quasar-schema = { version = "=0.1.0", path = "../schema" }
+heck.workspace = true
 quasar-idl-schema = { version = "=0.1.0", path = "schema" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
