--- schema/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "quasar-schema"
-description = "Shared schema types for Quasar interfaces"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-authors.workspace = true
-documentation = "https://docs.rs/quasar-schema/0.1.0"
-readme = "README.md"
-keywords = ["solana", "quasar", "schema", "code-generation"]
-categories = ["development-tools", "data-structures"]
-include = ["/src/**", "/README.md", "/LICENSE-APACHE", "/LICENSE-MIT"]
-
-[lints]
-workspace = true
-
-[lib]
-name = "quasar_schema"
-path = "src/lib.rs"
-doctest = false
