--- schema/LICENSE-APACHE
@@ -1,1 +1,0 @@
-../LICENSE-APACHE
