--- schema/README.md
@@ -1,10 +1,0 @@
-# quasar-schema
-
-Quasar's runtime, proc macros, IDL tooling, and CLI share these schema types.
-Most application code should depend on [`quasar-lang`](https://crates.io/crates/quasar-lang);
-this crate supports framework internals.
-
-- [Quasar guide](https://quasar-lang.com/docs)
-- [API documentation](https://docs.rs/quasar-schema/0.1.0)
-
-Licensed under Apache-2.0 or MIT, at your option.
